Refactor tips and comment action buttons until SSO part has been merged and released
