+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/exception/tips/TipsAdapter.java
@@ -1,0 +1,55 @@
+package it.niedermann.nextcloud.deck.ui.exception.tips;
+
+import android.content.Intent;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
+import androidx.core.util.Consumer;
+import androidx.recyclerview.widget.RecyclerView;
+
+import java.util.LinkedList;
+import java.util.List;
+
+import it.niedermann.nextcloud.deck.R;
+
+public class TipsAdapter extends RecyclerView.Adapter<TipsViewHolder> {
+
+    @NonNull
+    private Consumer<Intent> actionButtonClickedListener;
+    @NonNull
+    private List<TipsModel> tips = new LinkedList<>();
+
+    public TipsAdapter(@NonNull Consumer<Intent> actionButtonClickedListener) {
+        this.actionButtonClickedListener = actionButtonClickedListener;
+    }
+
+    @NonNull
+    @Override
+    public TipsViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        final View v = LayoutInflater.from(parent.getContext()).inflate(R.layout.item_tip, parent, false);
+        return new TipsViewHolder(v);
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull TipsViewHolder holder, int position) {
+        holder.bind(tips.get(position), actionButtonClickedListener);
+    }
+
+    @Override
+    public int getItemCount() {
+        return tips.size();
+    }
+
+    public void add(@StringRes int text) {
+        add(text, null);
+    }
+
+    public void add(@StringRes int text, @Nullable Intent primaryAction) {
+        tips.add(new TipsModel(text, primaryAction));
+        notifyItemInserted(tips.size());
+    }
+}
